+++ src/v0.8/shared/events/DatacapsEvents.sol
@@ -16,8 +16,5 @@ pragma solidity ^0.8.21;
 
 library DatacapsEvents {
     ///notice:  Event emitted when datacap is allocated to a matching
-    event DatacapAllocated(
-        uint256 indexed matchingId,
-        uint64 allocatedCapacity
-    );
+    event DatacapAllocated(uint64 indexed matchingId, uint64 allocatedCapacity);
 }
